0.2.0 업데이트 버그 수정 및 안정화 - UI 개선 - 각종 버그 픽스 기능 추가 - 참여 비율 제공 - 농활 소개 확장 - 후기 별점 기능 제공

+++ nongchung_iOS/VO/Review/MyReviewDataVO.swift
@@ -10,7 +10,7 @@ import Foundation
 
 struct MyReviewDataVO: Codable {
     let content: String?
-    let star: Double?
+    let star: Int?
     let img: [String]?
     let farmImg: String?
     let name: String?
